+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Added more keywords to match SLURM errors (\#843).
 * Fixed bug in sandbox pages (\#839);
 * Displayed all kinds of recent task activities in tasks management page. (\#835);
-* Managed resources and profiles (\#845):
+* Managed resources and profiles (\#845, \#862):
     * Added new admin pages for resources and profiles;
     * Updated admin user pages to associate users with profiles;
     * Updated admin settings page;